Refactor cite and dynamic cite pages
